requirements.txt
pandas
folium
psycopg2-binary
python-dotenv